+++ Invisi/backend/create-vapid.js
@@ -1,4 +1,0 @@
-// create-vapid.js
-const webpush = require('web-push');
-const vapidKeys = webpush.generateVAPIDKeys();
-console.log(vapidKeys);